Update README.md
Hey guys, there is a prototype of the timer based on existing sketches. Some links (like Play, Back, To Do) are clickable. Proposing to use this logo, as it looks like play button and clock at the same time. Not sure if the name was already chosen. Hoping to contribute during the testing stage. Thank you guys for doing your coding magic!